refactor platform bitmap factory
