+++ imagepipeline/src/main/java/com/facebook/imagepipeline/bitmaps/ArtBitmapFactory.java
@@ -36,33 +36,12 @@ import java.nio.ByteBuffer;
  */
 @TargetApi(Build.VERSION_CODES.LOLLIPOP)
 @ThreadSafe
-class ArtBitmapFactory extends PlatformBitmapFactory {
-
-  /**
-   * Size of temporary array. Value recommended by Android docs for decoding Bitmaps.
-   */
-  private static final int DECODE_BUFFER_SIZE = 16 * 1024;
+public class ArtBitmapFactory extends PlatformBitmapFactory {
 
   private final BitmapPool mBitmapPool;
 
-  /**
-   * ArtPlatformImageDecoder decodes images from InputStream - to do so we need to provide
-   * temporary buffer, otherwise framework will allocate one for us for each decode request
-   */
-  @VisibleForTesting final SynchronizedPool<ByteBuffer> mDecodeBuffers;
-
-
-  // TODO (5884402) - remove dependency on JfifUtil
-  private static final byte[] EOI_TAIL = new byte[] {
-      (byte) JfifUtil.MARKER_FIRST_BYTE,
-      (byte) JfifUtil.MARKER_EOI};
-
-  ArtBitmapFactory(BitmapPool bitmapPool, int maxNumThreads) {
+  public ArtBitmapFactory(BitmapPool bitmapPool) {
     mBitmapPool = bitmapPool;
-    mDecodeBuffers = new SynchronizedPool<>(maxNumThreads);
-    for (int i = 0; i < maxNumThreads; i++) {
-      mDecodeBuffers.release(ByteBuffer.allocate(DECODE_BUFFER_SIZE));
-    }
   }
 
   /**
@@ -83,135 +62,5 @@ class ArtBitmapFactory extends PlatformBitmapFactory {
     Bitmap bitmap = mBitmapPool.get(sizeInBytes);
     Bitmaps.reconfigureBitmap(bitmap, width, height, bitmapConfig);
     return CloseableReference.of(bitmap, mBitmapPool);
-  }
-
-  /**
-   * Creates a bitmap from encoded bytes.
-   * @param encodedImage the encoded image with a reference to the encoded bytes
-   * @param bitmapConfig the {@link android.graphics.Bitmap.Config}
-   * used to create the decoded Bitmap
-   * @return the bitmap
-   * @exception java.lang.OutOfMemoryError if the Bitmap cannot be allocated
-   */
-  @Override
-  public CloseableReference<Bitmap> decodeFromEncodedImage(
-      EncodedImage encodedImage,
-      Bitmap.Config bitmapConfig) {
-    final BitmapFactory.Options options = getDecodeOptionsForStream(encodedImage, bitmapConfig);
-    boolean retryOnFail=options.inPreferredConfig != Bitmap.Config.ARGB_8888;
-    try {
-      return decodeStaticImageFromStream(encodedImage.getInputStream(), options);
-    } catch (RuntimeException re) {
-      if (retryOnFail) {
-        return decodeFromEncodedImage(encodedImage, Bitmap.Config.ARGB_8888);
-      }
-      throw re;
-    }
-  }
-
-  /**
-   * Creates a bitmap from encoded JPEG bytes. Supports a partial JPEG image.
-   * @param encodedImage the encoded image with reference to the encoded bytes
-   * @param bitmapConfig the {@link android.graphics.Bitmap.Config}
-   * used to create the decoded Bitmap
-   * @param length the number of encoded bytes in the buffer
-   * @return the bitmap
-   * @exception java.lang.OutOfMemoryError if the Bitmap cannot be allocated
-   */
-  @Override
-  public CloseableReference<Bitmap> decodeJPEGFromEncodedImage(
-      EncodedImage encodedImage,
-      Bitmap.Config bitmapConfig,
-      int length) {
-    boolean isJpegComplete = encodedImage.isCompleteAt(length);
-    final BitmapFactory.Options options = getDecodeOptionsForStream(encodedImage, bitmapConfig);
-
-    InputStream jpegDataStream = encodedImage.getInputStream();
-    // At this point the InputStream from the encoded image should not be null since in the
-    // pipeline,this comes from a call stack where this was checked before. Also this method needs
-    // the InputStream to decode the image so this can't be null.
-    Preconditions.checkNotNull(jpegDataStream);
-    if (encodedImage.getSize() > length) {
-      jpegDataStream = new LimitedInputStream(jpegDataStream, length);
-    }
-    if (!isJpegComplete) {
-      jpegDataStream = new TailAppendingInputStream(jpegDataStream, EOI_TAIL);
-    }
-    boolean retryOnFail=options.inPreferredConfig != Bitmap.Config.ARGB_8888;
-    try {
-      return decodeStaticImageFromStream(jpegDataStream, options);
-    } catch (RuntimeException re) {
-      if (retryOnFail) {
-        return decodeFromEncodedImage(encodedImage, Bitmap.Config.ARGB_8888);
-      }
-      throw re;
-    }
-  }
-
-  @Override
-  protected boolean isPinBitmapEnabled() {
-    return false;
-  }
-
-  private CloseableReference<Bitmap> decodeStaticImageFromStream(
-      InputStream inputStream,
-      BitmapFactory.Options options) {
-    Preconditions.checkNotNull(inputStream);
-    int sizeInBytes = BitmapUtil.getSizeInByteForBitmap(
-        options.outWidth,
-        options.outHeight,
-        options.inPreferredConfig);
-    final Bitmap bitmapToReuse = mBitmapPool.get(sizeInBytes);
-    if (bitmapToReuse == null) {
-      throw new NullPointerException("BitmapPool.get returned null");
-    }
-    options.inBitmap = bitmapToReuse;
-
-    Bitmap decodedBitmap;
-    ByteBuffer byteBuffer = mDecodeBuffers.acquire();
-    if (byteBuffer == null) {
-      byteBuffer = ByteBuffer.allocate(DECODE_BUFFER_SIZE);
-    }
-    try {
-      options.inTempStorage = byteBuffer.array();
-      decodedBitmap = BitmapFactory.decodeStream(inputStream, null, options);
-    } catch (RuntimeException re) {
-      mBitmapPool.release(bitmapToReuse);
-      throw re;
-    } finally {
-      mDecodeBuffers.release(byteBuffer);
-    }
-
-    if (bitmapToReuse != decodedBitmap) {
-      mBitmapPool.release(bitmapToReuse);
-      decodedBitmap.recycle();
-      throw new IllegalStateException();
-    }
-
-    return CloseableReference.of(decodedBitmap, mBitmapPool);
-  }
-
-  /**
-   * Options returned by this method are configured with mDecodeBuffer which is GuardedBy("this")
-   */
-  private static BitmapFactory.Options getDecodeOptionsForStream(
-      EncodedImage encodedImage,
-      Bitmap.Config bitmapConfig) {
-    final BitmapFactory.Options options = new BitmapFactory.Options();
-    // Sample size should ONLY be different than 1 when downsampling is enabled in the pipeline
-    options.inSampleSize = encodedImage.getSampleSize();
-    options.inJustDecodeBounds = true;
-    // fill outWidth and outHeight
-    BitmapFactory.decodeStream(encodedImage.getInputStream(), null, options);
-    if (options.outWidth == -1 || options.outHeight == -1) {
-      throw new IllegalArgumentException();
-    }
-
-    options.inJustDecodeBounds = false;
-    options.inDither = true;
-    options.inPreferredConfig = bitmapConfig;
-    options.inMutable = true;
-
-    return options;
   }
 }
