+++ imagepipeline/src/main/java/com/facebook/imagepipeline/bitmaps/HoneycombBitmapFactory.java
@@ -13,29 +13,29 @@ import javax.annotation.concurrent.ThreadSafe;
 
 import android.annotation.TargetApi;
 import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.graphics.Color;
 import android.os.Build;
 
 import com.facebook.common.references.CloseableReference;
 import com.facebook.imageformat.ImageFormat;
 import com.facebook.imagepipeline.image.EncodedImage;
-import com.facebook.imagepipeline.memory.FlexByteArrayPool;
 import com.facebook.imagepipeline.memory.PooledByteBuffer;
+import com.facebook.imagepipeline.platform.PlatformDecoder;
 
 /**
- * Factory implementation for Honeycomb
+ * Factory implementation for Honeycomb through Kitkat
  */
 @TargetApi(Build.VERSION_CODES.HONEYCOMB)
 @ThreadSafe
-public class HoneycombBitmapFactory extends DalvikBitmapFactory {
+public class HoneycombBitmapFactory extends PlatformBitmapFactory {
 
   private final EmptyJpegGenerator mJpegGenerator;
+  private final PlatformDecoder mPurgeableDecoder;
 
-  HoneycombBitmapFactory(EmptyJpegGenerator jpegGenerator,
-      FlexByteArrayPool flexByteArrayPool) {
-    super(flexByteArrayPool);
-    this.mJpegGenerator = jpegGenerator;
+  public HoneycombBitmapFactory(EmptyJpegGenerator jpegGenerator,
+                         PlatformDecoder purgeableDecoder) {
+    mJpegGenerator = jpegGenerator;
+    mPurgeableDecoder = purgeableDecoder;
   }
 
   /**
@@ -61,8 +61,8 @@ public class HoneycombBitmapFactory extends DalvikBitmapFactory {
       EncodedImage encodedImage = new EncodedImage(jpgRef);
       encodedImage.setImageFormat(ImageFormat.JPEG);
       try {
-        CloseableReference<Bitmap> bitmapRef =
-            decodeJPEGFromEncodedImage(encodedImage, bitmapConfig, jpgRef.get().size());
+        CloseableReference<Bitmap> bitmapRef = mPurgeableDecoder.decodeJPEGFromEncodedImage(
+            encodedImage, bitmapConfig, jpgRef.get().size());
         bitmapRef.get().eraseColor(Color.TRANSPARENT);
         return bitmapRef;
       } finally {
@@ -72,40 +72,4 @@ public class HoneycombBitmapFactory extends DalvikBitmapFactory {
       jpgRef.close();
     }
   }
-
-  @Override
-  protected boolean isPinBitmapEnabled() {
-    return true;
-  }
-
-  /**
-   * Decodes a byteArray into a purgeable bitmap
-   *
-   * @param bytesRef the byte buffer that contains the encoded bytes
-   * @return
-   */
-  @Override
-  protected Bitmap decodeByteArrayAsPurgeable(
-      CloseableReference<PooledByteBuffer> bytesRef,
-      BitmapFactory.Options options) {
-    return decodeFileDescriptorAsPurgeable(bytesRef, bytesRef.get().size(), null, options);
-  }
-
-  /**
-   * Decodes a byteArray containing jpeg encoded bytes into a purgeable bitmap
-   *
-   * <p> Adds a JFIF End-Of-Image marker if needed before decoding.
-   *
-   * @param bytesRef the byte buffer that contains the encoded bytes
-   * @return
-   */
-  @Override
-  protected Bitmap decodeJPEGByteArrayAsPurgeable(
-      CloseableReference<PooledByteBuffer> bytesRef,
-      int length,
-      BitmapFactory.Options options) {
-    byte[] suffix = endsWithEOI(bytesRef, length) ? null : EOI;
-    return decodeFileDescriptorAsPurgeable(bytesRef, length, suffix, options);
-  }
-
 }
